Add permissions and update tests

+++ store/product/views.py
@@ -3,6 +3,7 @@ from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from .permissions import IsAdmin
 from .models import Product
 from .serializers import ProductSerializer
 
@@ -14,7 +15,8 @@ class ProductListView(APIView):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def post(self, request):
-        # TODO Only admin can create a product
+        if not IsAdmin().has_permission(request, request.user):
+            return Response({"detail":"Only admin can add a new product"}, status=status.HTTP_403_FORBIDDEN)
         serializer = ProductSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -28,7 +30,8 @@ class ProductDetailView(APIView):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def patch(self, request, id):
-        # TODO Only admin can edit a product
+        if not IsAdmin().has_permission(request, request.user):
+            return Response({"detail":"Only admin can modify a new product"}, status=status.HTTP_403_FORBIDDEN)
         product = get_object_or_404(Product, id=id)
         serializer = ProductSerializer(product, data=request.data, partial=True)
         if serializer.is_valid():
@@ -37,7 +40,8 @@ class ProductDetailView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, id):
-        # TODO Only admin can delete a product
+        if not IsAdmin().has_permission(request, request.user):
+            return Response({"detail":"Only admin can delete a new product"}, status=status.HTTP_403_FORBIDDEN)
         product = get_object_or_404(Product, id=id)
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
